+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## [0.7.0]
+
+- Library now also extends `String` so there is no need of using the bang operator on `String`s that are not `null`
+- Bumps dependencies
+
+## [0.6.14] - Code clean up
+
+- Removes Flutter from dependencies
+- Improves documentation
+
+## [0.6.13] - New Extensions
+
+- `nullIf(String? comparisonString)` : Returns `null` if `String` equals `comparisonString`. Otherwise return the `String` as is.
+- `emptyIf(String? comparisonString)` : Returns empty `String` if it equalts to `comparisonString`. Otherwise return the `String` as is.
+- `replaceFirst(String pattern, String replacement)` : Returns a new string with the first occurrence of the given pattern replaced with the replacement string.
+- `replaceLast(String pattern, String replacement)` : Returns a new string with the last occurrence of the given pattern replaced with the replacement string.
+
 ## [0.6.12] - New Extensions
 
 - `insertAt` : Inserts a `String` at a specified index